+++ examples/basic-create-react-app/src/components/App.js
@@ -9,7 +9,7 @@ function App(props) {
     return <div className="pure-g">
         <div className="pure-u-1-3">
             <ApiVersion apiVersion={props.apiVersion}/>
-            <Settings settings={props.context.global.settings} />
+            <Settings settings={props.context.general.settings} />
         </div>
         <div className="pure-u-2-3">
             <Projects projects={props.context.projects} />